requirements.txt
appdirs==1.4.4
asgiref==3.5.2
astroid==2.12.10
async-generator==1.10
attrs==22.1.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.0
beautifulsoup4==4.11.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
coverage==6.4.4
cryptography==38.0.1
dill==0.3.5.1
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field[phonenumberslite]==7.0.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
flake8==5.0.4
h11==0.13.0
idna==3.4
imgkit==1.2.2
isodate==0.6.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
msrest==0.7.1
oauthlib==3.2.2
outcome==1.2.0
phonenumberslite==8.12.56
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.0
pyflakes==2.5.0
pylint==2.15.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
PySocks==1.7.1
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.4.3
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starkbank-ecdsa==2.1.0
tomli==2.0.1
tomlkit==0.11.4
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
veryfi==3.2.0
webdriver-manager==3.8.3
webdrivermanager==0.10.0
wrapt==1.14.1
wsproto==1.2.0
drf-spectacular==0.25.1
